refactor: Move tillCount logic to custom hook
Moved the tillCount state management logic from a state file to a custom hook. This improves code organization and reusability.

- Moved tillCount.state.ts to useTillCount.ts
- Updated import statements in App.tsx to reflect the change.
